compile with a local version of 3c

+++ rollingarray.h
@@ -23,7 +23,7 @@ ra_item_t rollingarray_get_item(rollingarray_t *ra, size_t ind);
 ra_item_t rollingarray_last_item(rollingarray_t *ra);
 ra_item_t rollingarray_get_item_backwards(rollingarray_t *ra, size_t ind);
 void rollingarray_add_item(rollingarray_t *ra, ra_item_t val);
-int rollingarray_size(rollingarray_t *ra);
+size_t rollingarray_size(rollingarray_t *ra);
 size_t rollingarray_to_array(rollingarray_t *ra, ra_item_t arr[],
                              bool reversed);
 bool rollingarray_has_item(rollingarray_t *ra, ra_item_t item);
